Trying to write to a DB
Initially, I manually input the details into the database and then checked if the details existed or not. This worked, however it is not at all convenient or helpful, so today I tried to allow the user to register their own details into the database. However, when I tried this, an exception was thrown and I have yet troubleshoot this issue