Writing to and reading from the DB
Since the last Git push, I fixed the last exception that was thrown when trying to write to the database. The issue was that the ID table had "Not null" ticked in the Database, meaning it doesn't allow nulls. I was ecstatic when I successfully wrote to the database, however this happiness was short lived as when I try to check if a user already exists my application, I enter an eternal loop, I looked around for answers on logic but couldn't find any that works if (C# WPF DB Browser for SQLite). This is when I remembered that my good friend Ethan worked with C# interfaces last year, so I asked him about it and I ended up not needing the loop. However, now my application freezes when I try write to the DB and no exception is thrown which means I don't even know where to start when it comes to troubleshooting.

A side note: I initially decided to do a push each week, however I have started to do one whenever I reach a milestone or roadblock. I also believe that from now on, my Git pushes will have similar length explanations to this one.