The hard part of my project
I have almost completed the UI of the page where the user can add and see workouts. This is when I began to look at how to create a new table for each user when their workouts would be logged, then I realised I have no idea how to do this. I knew that I would need to link the table with foreign keys and primary keys of the user however, I was not sure if it was possible through C# commands. I did a little bit of research but could not find an answer. So I did a bit of designing of the settings/edit profile page. This week I slowed down in progress due to   being busy with other subjects.