Saving the users Id
After talking to Mr. Fong, I then created a table called 'Workout' and the process in which the user saves a workout to the table requires the user primary key to make the field added unique to that user. I did hours of research before succumbing to Stack overflow and getting correct answer in a matter of minutes. I have now saved the UserID in a variable called 'primaryKey' and started to design the list where the workouts will be displayed.