Workout Page completed
I binded the buttons to add a workout to the backend functions and was able to create workouts in the database. I also created the listview on the GUI and have managed to display only the workouts relating to the user that is logged in. By searching for their UserID. I also customised the design of the list view to match the concurrent theme throughout the software and saved the date of each workout as well. To conclude the page I also display the total time of workouts the user has done, however, I tried to convert the minutes into hours, and spent 70 minutes trying various methods to keep the hours to two decimal places so that it wasn't recurring. In the end, I changed the total back to minutes and finished the Workouts page