Settings page complete
Once I finished the workout page, I started working on the settings page, in order for the user to view their current details, I saved all the personal details of the logged in user into global variables to allow them to change their details. I also made it so the user could change their personal details in the database. Getting this done was difficult because I was still learning both XAML and the Entity framework at the time. I then incorporated a refresh button so that the user knows that their details has been updated. With that final feature, I had finished the settings page.

I looked into creating a chat/messaging system, but with the time restraints, I decided against it and  started looking features that I gave up early on because of their time consumption.

These are the features that I might add if their is time:

Make a separate button theme for close button and MessageBox

Store passwords with salted hashes

Column separators in the workout table

Convert the minutes into hours and display as a decimal on the workout screen